Refactor EmbeddedNavController: Remove SnapKit and Organize InitialViewController into Multiple Files
